+++ changelog.md
@@ -1,15 +1,26 @@
 # Versions
 
-### v0.87-beta
+### v0.9.0-beta
+
+- added time progress bar to display current video duration
+- added start time / end time
+- added skip buttons to jump +/- 5 seconds
+- added support to copy current timestamp into clipboard
+
+### v0.8.1-beta
+
+- added default resolution for Linux system (now app supports Win, MacOS, and Linux)
+
+### v0.8.0-beta
 
 - application now can render up to 4 videos in preview
 - adjusted app resolution to ensure video ratio is unchanged in preview
 
-### v0.85-beta
+### v0.7.0-beta
 
 - application now supports up to 4 video tracks + 4 audio tracks
 - application can generate video splicing using timeline text
 
-### v0.80-alpha
+### v0.6.0-alpha
 
 - application now can only render 2 videos simutaneously